lsp: move wasm module to file
